Initial commit: Walking skeleton
This initial commit setup a simple provisioning using Vagrant and Shell
as well the project structure with a failing test for the first feature.

All provisioning scripts are located in the provision folder. There are
two machines being provisioned: prestashop and presta_bridge. The first one
is a PrestaShop installation and the latter one is a Rails app that will
consume the PrestaShop API.

For both cases the main script, which triggers the whole installation,
is named setup.sh and both of them includes more scripts to install or
configure a specific binary or library.

For prestashop we have the following provisioning scripts:

* apache.sh: Configures Apache Web Server to serve PrestaShop
* composer.sh: Install Composer following the instructions in the
               website
* mysql.sh: Installs MySQL binary and defines root password as "abc123"
* php.sh: Install PHP 5 and all extensions to run PrestaShop
* prestashop.sh: This script automates the installation of the
                 PrestaShop store avoiding any interaction with a web
                 browser

And for presta_bridge machine we have the following scripts:

* bundler.sh: Which installs bundler gem and run bundle install
* postgresql.sh: Which installs PostgreSQL 9.6 packages and creates a
				 user for vagrant
* rvm.sh: Which installs and configure RVM with Ruby 2.4.0

The two VMs are instances of Ubuntu 14.04 LTS instance because of the ease of
installing and configuring packages. Also, both machines will use 1GB of
memory.

For presta_bridge (Rails app) NFS to speed up I/O writes so your
password will be required.

To access presta_bridge on your host machine go to http://localhost:8081
and run all rails related command on the guest machine (rails c, rake
db:migrate, etc).

The PrestaShop installation is available on http://10.0.0.10 and to
access the admin panel use the following credentials:

username: pub@prestashop.com
password: 0123456789

The failing test in users controller is the start point of the
application it will be fixed in the next iterations and other features
as well.