Create user feature
This commit enables the creation of users in the system. It uses Rails'
resources feature to add the needed routes (even if some of the
generated routes are not proper handled).

A normal user has the following attributes: name, email and password.
The password is crypted using bcrypt by Rails to avoid storing plain
text password.

The new form uses Bootstrap framework because it's the most familiar
one to me.

Finally, in case of any error when creating a new user a flash message
will be displayed showing the errors to the user and redirecting it to
the form again.