Create user's settings page
The new Settings page is where user can configure a store and see its
products.

After the creation of a new user it'll be redirected to the Settings
page the same will happen when the user is successfully authenticated.

To show store information a new table was created: stores. This table
contains the following fields:

* user_id: Reference to users table (store owner)
* name: Store name
* url: A public accessible URL pointing to the store (this information
       will be used to consume its API)
* api_key: The key generated by PrestaShop to consume its API